Struct google_replicapool1_beta2::InstanceGroupManager

pub struct InstanceGroupManager {
    pub target_size: Option<i32>,
    pub group: Option<String>,
    pub description: Option<String>,
    pub auto_healing_policies: Option<Vec<ReplicaPoolAutoHealingPolicy>>,
    pub fingerprint: Option<String>,
    pub kind: Option<String>,
    pub base_instance_name: Option<String>,
    pub target_pools: Option<Vec<String>>,
    pub instance_template: Option<String>,
    pub current_size: Option<i32>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

An Instance Group Manager resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set target pools instance group managers (none)

list instance group managers (none)

insert instance group managers (request)

get instance group managers (response)

abandon instances instance group managers (none)

recreate instances instance group managers (none)

delete instance group managers (none)

set instance template instance group managers (none)

resize instance group managers (none)

delete instances instance group managers (none)

Fields

target_size

[Output only] The number of instances that the manager is attempting to maintain. Deleting or abandoning instances affects this number, as does resizing the group.

group

[Output only] The full URL of the instance group created by the manager. This group contains all of the instances being managed, and cannot contain non-managed instances.

description

An optional textual description of the instance group manager.

auto_healing_policies

The autohealing policy for this managed instance group. You can specify only one value.

fingerprint

[Output only] Fingerprint of the instance group manager. This field is used for optimistic locking. An up-to-date fingerprint must be provided in order to modify the Instance Group Manager resource.

kind

[Output only] The resource type. Always replicapool#instanceGroupManager.

base_instance_name

The base instance name to use for instances in this group. The value must be a valid RFC1035 name. Supported characters are lowercase letters, numbers, and hyphens (-). Instances are named by appending a hyphen and a random four-character string to the base instance name.

target_pools

The full URL of all target pools to which new instances in the group are added. Updating the target pool values does not affect existing instances.

instance_template

The full URL to an instance template from which all new instances will be created.

current_size

[Output only] The number of instances that currently exist and are a part of this group. This includes instances that are starting but are not yet RUNNING, and instances that are in the process of being deleted or abandoned.

creation_timestamp

[Output only] The time the instance group manager was created, in RFC3339 text format.

id

[Output only] A server-assigned unique identifier for the resource.

self_link

[Output only] The fully qualified URL for this resource.

name

The name of the instance group manager. Must be 1-63 characters long and comply with RFC1035. Supported characters include lowercase letters, numbers, and hyphens.