Crate google_replicapool1_beta2

This documentation was generated from replicapool crate version 0.1.15+20160512, where 20160512 is the exact revision of the replicapool:v1beta2 schema built by the mako code generator v0.1.15.

Everything else about the replicapool v1_beta2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

instance group managers

abandon instances, delete, delete instances, get, insert, list, recreate instances, resize, set instance template and set target pools

zone operations

get and list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.instance_group_managers().set_target_pools(...).doit()
let r = hub.instance_group_managers().list(...).doit()
let r = hub.instance_group_managers().insert(...).doit()
let r = hub.instance_group_managers().get(...).doit()
let r = hub.instance_group_managers().abandon_instances(...).doit()
let r = hub.instance_group_managers().recreate_instances(...).doit()
let r = hub.instance_group_managers().delete(...).doit()
let r = hub.instance_group_managers().set_instance_template(...).doit()
let r = hub.instance_group_managers().resize(...).doit()
let r = hub.instance_group_managers().delete_instances(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-replicapool1_beta2 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_replicapool1_beta2 as replicapool1_beta2;
use replicapool1_beta2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use replicapool1_beta2::Replicapool;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Replicapool::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instance_group_managers().list("project", "zone")
             .page_token("sanctus")
             .max_results(79)
             .filter("amet")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

InstanceGroupManager
An Instance Group Manager resource.

InstanceGroupManagerAbandonInstanceCall
Removes the specified instances from the managed instance group, and from any target pools of which they were members, without deleting the instances.

InstanceGroupManagerDeleteCall
Deletes the instance group manager and all instances contained within. If you'd like to delete the manager without deleting the instances, you must first abandon the instances to remove them from the group.

InstanceGroupManagerDeleteInstanceCall
Deletes the specified instances. The instances are deleted, then removed from the instance group and any target pools of which they were a member. The targetSize of the instance group manager is reduced by the number of instances deleted.

InstanceGroupManagerGetCall
Returns the specified Instance Group Manager resource.

InstanceGroupManagerInsertCall
Creates an instance group manager, as well as the instance group and the specified number of instances.

InstanceGroupManagerList
There is no detailed description.

InstanceGroupManagerListCall
Retrieves the list of Instance Group Manager resources contained within the specified zone.

InstanceGroupManagerMethods
A builder providing access to all methods supported on instanceGroupManager resources. It is not used directly, but through the Replicapool hub.

InstanceGroupManagerRecreateInstanceCall
Recreates the specified instances. The instances are deleted, then recreated using the instance group manager's current instance template.

InstanceGroupManagerResizeCall
Resizes the managed instance group up or down. If resized up, new instances are created using the current instance template. If resized down, instances are removed in the order outlined in Resizing a managed instance group.

InstanceGroupManagerSetInstanceTemplateCall
Sets the instance template to use when creating new instances in this group. Existing instances are not affected.

InstanceGroupManagerSetTargetPoolCall
Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.

InstanceGroupManagersAbandonInstancesRequest
There is no detailed description.

InstanceGroupManagersDeleteInstancesRequest
There is no detailed description.

InstanceGroupManagersRecreateInstancesRequest
There is no detailed description.

InstanceGroupManagersSetInstanceTemplateRequest
There is no detailed description.

InstanceGroupManagersSetTargetPoolsRequest
There is no detailed description.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Operation
An operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationList
There is no detailed description.

OperationWarnings
[Output Only] If there are issues with this operation, a warning is returned.

OperationWarningsData
[Output only] Metadata for this warning in key:value format.

ReplicaPoolAutoHealingPolicy
There is no detailed description.

Replicapool
Central instance to access all Replicapool related resource activities

ZoneOperationGetCall
Retrieves the specified zone-specific operation resource.

ZoneOperationListCall
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethods
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Replicapool hub.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.